backend/requirements.txt
# ─── Web Framework ───────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# ─── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── AI / LLM ─────────────────────────────────────────────────────────────────
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.7
langchain-openai==0.2.6
openai==1.54.4

# ─── Embeddings & Vector Store ───────────────────────────────────────────────
faiss-cpu==1.9.0
sentence-transformers==3.2.1

# ─── Document Loading ─────────────────────────────────────────────────────────
pypdf==5.0.1
python-docx==1.1.2

# ─── Config ───────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# ─── HTTP / Async ─────────────────────────────────────────────────────────────
httpx==0.27.2
aiofiles==24.1.0

# ─── AWS (optional — for image upload) ───────────────────────────────────────
boto3==1.35.44
